Management 모듈 리팩토링  (#143)
- Management 모듈 리팩토링
- Pub/Sub (Kafka) 를 통해 다중 트랜잭션 처리 로직 생성
- 다중 트랜잭션에 대한 롤백 로직은 미작성
- 추후 안드로이드 테스트 진행 후, 롤백 처리에 관한 로직 생성 에정

+++ domains/play-mong-domain/src/main/java/com/mongs/play/domain/mong/enums/MongExp.java
@@ -6,11 +6,11 @@ import lombok.Getter;
 @AllArgsConstructor
 public enum MongExp {
 
-    STROKE(5),
-    ATTENDANCE(5),
-    CLEANING_POOP(1),
-    EAT_THE_FOOD(10)
+    STROKE(5D),
+    ATTENDANCE(5D),
+    CLEANING_POOP(1D),
+    EAT_THE_FOOD(10D)
     ;
 
-    public final Integer exp;
+    public final Double exp;
 }
